Build exports jar

+++ src/com/redpois0n/Build.java
@@ -4,18 +4,25 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.io.FileReader;
 import java.io.FileWriter;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.io.PrintWriter;
-import java.nio.channels.FileChannel;
 
 public class Build {
 
-	public static final String infoFile = "files/Info.plist";
-	public static final String launcherFile = "files/launcher";
+	public static final String runpath;
+
+	static {
+		runpath = Main.class.getProtectionDomain().getCodeSource().getLocation().getPath();
+	}
+
+	public static final String infoFile = "/files/Info.plist";
+	public static final String launcherFile = "/files/launcher.txt";
 
 	public static void build(File input, File output, File icon, String mainclass, String appTitle, boolean dockIcon, int minimumVersion) throws Exception {
-		File basePath = new File(output.getAbsolutePath() + ".app", "/Contents");
+		File basePath = new File(output.getAbsolutePath(), "/Contents");
 
 		new File(basePath, "/MacOS").mkdirs();
 		new File(basePath, "/Resources").mkdirs();
@@ -28,44 +35,51 @@ public class Build {
 
 		launcher.createNewFile();
 
-		if (icon.exists()) {
+		if (icon != null && icon.exists()) {
 			copyFile(icon, iconFile);
 		}
-		
+
 		copyLauncher(launcher, input.getName(), appTitle, dockIcon, minimumVersion);
-	    
-	    copyFile(new File(infoFile), info);
-	    copyFile(input, jar);
-	    
-	    String chmod = "chmod +x " + launcher.getAbsolutePath();
-	    Runtime.getRuntime().exec(new String[] { "/bin/bash", "-c", chmod });
-	    
+
+		copy(Main.class.getResourceAsStream(infoFile), new FileOutputStream(info));
+		copyFile(input, jar);
+
+		// String chmod = "chmod +x " + launcher.getAbsolutePath();
+		// Runtime.getRuntime().exec(new String[] { "/bin/bash", "-c", chmod });
+
 	}
 
 	public static void copyFile(File input, File output) throws Exception {
 		if (!output.getParentFile().exists()) {
 			output.getParentFile().mkdirs();
 		}
-		FileChannel source = new FileInputStream(input).getChannel();
-		FileChannel destination = new FileOutputStream(output).getChannel();
-		destination.transferFrom(source, 0L, source.size());
 
-		source.close();
+		copy(new FileInputStream(input), new FileOutputStream(output));
+	}
 
-		destination.close();
+	public static long copy(InputStream input, OutputStream output) throws Exception {
+		byte[] buffer = new byte[1024];
+		int n = 0;
+		long count = 0L;
+		while ((n = input.read(buffer)) != -1) {
+			output.write(buffer, 0, n);
+			count += n;
+		}
 
+		input.close();
+		output.close();
+		return count;
 	}
 
 	public static void copyLauncher(File launcherDest, String jarName, String appTitle, boolean showIcon, int minimumVersion) throws Exception {
-
-		FileReader orig = new FileReader(launcherFile);
+		InputStream is = Main.class.getResourceAsStream(launcherFile);
 
 		FileWriter dest = new FileWriter(launcherDest.getAbsolutePath());
 
 		BufferedReader br = null;
 		PrintWriter pw = null;
 
-		br = new BufferedReader(orig);
+		br = new BufferedReader(new InputStreamReader(is));
 		pw = new PrintWriter(dest);
 
 		int nline = 1;
